Add a summarize_screens script (#12)
Adds a script to produce CSVs that summarize all of the .screen files in a directory.

Intended to supercede flag for the internal use case of for summarizing output files while caring a lot about the outcomes of the different pipeline steps ("nucleotide vs. protein screen") as opposed to an interpretable recommendation ("FLAG: match at bases X to Y to regulated virus Z") or ("INFO: mix of regulated and non-regulated pathogens").