Provide Diamond with --jobs, not "--processes" (#14)
When investigating the inconsistent `--processes` tag, I realised that `run_diamond.sh` was not written in a robust way, and "processes" is not a good description of its function.

Specifically, in this command (pulled out the parts not relevant to parallelization):

```bash
ls ${DB_PATH}/nr*.dmnd | parallel -j ${PROCESSES}
diamond blastx --quiet -d ${DB_PATH}/nr.{%}.dmnd \
--threads ${THREADS} -o ${OUTPUT}.{%}.tsv
```

* `PROCESSES` is not really "processes", it's the number of jobs we tell parallel to run at once
* we use the `{%}` replacement string, which is the job number, to decide which database we use in `--d`... I think they way we've set this up we need to literally always be counting from 1-6. (which is what `screen.py` was doing, but not `run_pipeline.sh`). Based on some local testing of parallel, if we were running with only 2 processes, then the outputs would get (I think? not sure what diamond does with pre-existing output files) overwritten

I have changed this to set the number of jobs to run in parallel based on the threads and CPUs (unless a number is supplied) and made some other parts of run_diamond a bit more robust, then propagated those changes.