Initial integration of databases with screen.py, untested.
Modify the script to output whether the flag was for virus, bacteria, or eukaryote

Standardised the unit tests for JSON IO handler, split into the testing directory, and fixed bug regarding the check_biorisk == literal, rather than using is.

Updated to a almost functional version of the refactor, needs systematic testing of new database functionality.